+++ config/config.exs
@@ -1,0 +1,2 @@
+use Mix.Config
+import_config "../apps/*/config/config.exs"
